FIX: Returnerer 403 dersom veileder prøver å fjerne en deltakelse der deltaker har søkt (#73)

+++ app/src/main/kotlin/no/nav/ung/deltakelseopplyser/domene/register/UngdomsprogramregisterService.kt
@@ -68,6 +68,18 @@ class UngdomsprogramregisterService(
         }
 
         val ungdomsprogramDAO = forsikreEksistererIProgram(id)
+
+        if (ungdomsprogramDAO.harSøkt) {
+            logger.error("Klarte ikke å slette deltaker fra programmet med id $id, fordi deltakeren har søkt")
+            throw ErrorResponseException(
+                HttpStatus.FORBIDDEN,
+                ProblemDetail.forStatus(HttpStatus.FORBIDDEN).also {
+                    it.detail = "Deltakeren har søkt og deltakelsen kan derfor ikke slettes"
+                },
+                null
+            )
+        }
+
         deltakelseRepository.delete(ungdomsprogramDAO)
 
         if (deltakelseRepository.existsById(id)) {
